Updated tmux copy and paste bindings. And edited TERM colours.
copy and paste bindings should now match vim, this will need to be
tested on a different terminal. And, copying to /dev/clipboard works on
git bash but will need to make sure it works on linux.

TERM colours have been updated to screen but might need to change to
xterm on different terminal. If this is required will most likely have
to have local settings for TERM colours.